requirements.txt
annotated-types==0.7.0
anyio==4.9.0
apturl==0.5.2
asgiref==3.8.1
bcrypt==3.2.0
blinker==1.4
Brlapi==0.8.3
CacheControl==0.14.2
cachetools==5.5.2
certifi==2020.6.20
chardet==4.0.0
charset-normalizer==3.4.1
click==8.0.3
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
defusedxml==0.7.1
distro==1.7.0
distro-info==1.1+ubuntu0.2
dj-database-url==2.3.0
Django==5.2
django-allauth==65.7.0
django-cors-headers==4.7.0
django-environ==0.12.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dnspython==2.7.0
duplicity==0.8.21
exceptiongroup==1.2.2
fasteners==0.14.1
firebase-admin==6.7.0
future==0.18.2
gcloud==0.18.3
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.20.1
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.69.2
grpcio==1.71.0
grpcio-status==1.71.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.20.2
httpx==0.28.1
idna==3.3
importlib-metadata==4.6.4
jeepney==0.7.1
jiter==0.9.0
jwcrypto==1.5.6
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
mongoengine==0.29.1
monotonic==1.6
more-itertools==8.10.0
msgpack==1.1.0
netifaces==0.11.0
oauth2client==4.1.3
oauthlib==3.2.0
olefile==0.46
openai==1.71.0
packaging==24.2
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
proto-plus==1.26.1
protobuf==5.29.4
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycairo==1.20.1
pycryptodome==3.22.0
pycups==2.0.1
pydantic==2.11.3
pydantic_core==2.33.1
PyGObject==3.42.1
PyJWT==2.9.0
pymacaroons==0.13.0
pymongo==3.11.4
PyNaCl==1.5.0
pyparsing==2.4.7
Pyrebase4==4.8.0
pyRFC3339==1.1
python-apt==2.4.0+ubuntu4
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
python-decouple==3.8
python-dotenv==1.1.0
python-jwt==4.1.0
python3-openid==3.2.0
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==0.10.1
rsa==4.9
SecretStorage==3.3.1
six==1.16.0
sniffio==1.3.1
social-auth-app-django==5.4.3
social-auth-app-django-mongoengine==1.0.0
social-auth-core==4.5.6
social-auth-storage-mongoengine==1.0.1
speedtest-cli==2.1.3
sqlparse==0.5.3
systemd-python==234
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.12.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.26.5
usb-creator==0.3.7
wadllib==1.3.6
whitenoise==6.9.0
xdg==5
xkit==0.0.0
zipp==1.0.0
